requirements.txt
aiohttp==3.7.4.post0
aiohttp-sse-client==0.2.0
alembic==1.5.8
aniso8601==9.0.1
appdirs==1.4.4
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
automium==0.2.6
automium-web==0.1.1
blinker==1.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
crc16==0.1.1
cryptography==3.4.7
distlib==0.3.1
Django==3.1.7
django-rest-auth-bearer==0.1.1
djangorestframework==3.12.4
ed25519==1.5
filelock==3.0.12
greenlet==1.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jsondiff==1.2.0
legacy==0.1.6
lxml==4.6.3
Mako==1.1.4
MarkupSafe==1.1.1
mnemonic==0.19
multidict==5.1.0
mysqlclient==2.0.3
numpy==1.20.3
pbkdf2==1.3
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
PyNaCl==1.4.0
python-dateutil==2.8.1
python-editor==1.0.4
pytils==0.3
pytz==2021.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.4.7
sqlparse==0.4.1
sseclient==0.0.27
stellar-base==1.1.2.0
stellar-base-sseclient==0.0.21
stellar-sdk==3.1.4
toml==0.10.2
tools==0.1.9
typing-extensions==3.7.4.3
urllib3==1.26.3
virtualenv==20.4.3
Werkzeug==1.0.1
wfastcgi==3.0.0
yarl==1.6.3

rpyc~=5.0.1
pyjsonrpc~=0.0.2
qrcode~=7.3
PyQRCode~=1.2.1
Pillow~=8.3.2
daemon~=1.2
phonenumbers~=8.12.45
environ~=1.0